requirements.txt
fastapi
uvicorn
requests
python-dotenvgit status